fix(rev-254): audit bugs — IF_OPENOVERLAY anim reset; MAP_PROJANIM *4; gouraud slope wrap
- client-08-02: IF_OPENOVERLAY (opcode 85) was missing the Java
  resetInterfaceAnimation(com) call for com >= 0 (Client.java:7279-7288
  @2e62978; a 254 delta) — animated overlay components rendered from
  stale seqFrame/seqCycle. Stale 245.2 cite corrected.
- client-08-04/wire-02: zone MAP_PROJANIM (opcode 37) dropped the
  Java `* 4` on both g1 height reads (Client.java:7682-7683 @2e62978;
  254 delta), rendering every projectile at 1/4 vertical offset.
- pix3d-B-01: GouraudRaster colour-slope multiply
  (arg7-arg6)*DivTable[arg3] now wraps to int32 before >>15, matching
  Java Pix3D.java:867 32-bit semantics (analog of the fixed
  textureRaster site at pix3d.go:2067).

Gates: build, vet, test -race, gofmt, wasm build, golangci-lint 0.

+++ pkg/jagex2/graphics/pix3d/pix3d.go
@@ -882,7 +882,9 @@ func GouraudRaster(arg0 []int, arg1, arg4, arg5, arg6, arg7 int) {
 			arg1 += arg4
 			arg3 = (arg5 - arg4) >> 2
 			if arg3 > 0 {
-				var8 = ((arg7 - arg6) * DivTable[arg3]) >> 15
+				// Java: Pix3D.java:867 — 32-bit product wraps before the >>15
+				// (overflows on steep colour deltas over short spans; audit pix3d-B-01)
+				var8 = int(int32((arg7-arg6)*DivTable[arg3])) >> 15
 			} else {
 				var8 = 0
 			}
